+++ src/server/HttpServerFactory.ts
@@ -4,5 +4,5 @@ import type { Server } from 'http';
  * A factory for HTTP servers
  */
 export interface HttpServerFactory {
-  startServer: (port: number) => Server;
+  startServer: (port: number | string) => Server;
 }