+++ test/unit/init/ServerInitializer.test.ts
@@ -14,21 +14,29 @@ describe('ServerInitializer', (): void => {
     serverFactory = {
       startServer: jest.fn().mockReturnValue(server),
     };
-    initializer = new ServerInitializer(serverFactory, 3000);
   });
 
   it('starts an HTTP server.', async(): Promise<void> => {
+    initializer = new ServerInitializer(serverFactory, 3000);
     await initializer.handle();
     expect(serverFactory.startServer).toHaveBeenCalledWith(3000);
   });
 
+  it('starts a server on the specified Unix Domain Socket.', async(): Promise<void> => {
+    initializer = new ServerInitializer(serverFactory, undefined, '/tmp/css.sock');
+    await initializer.handle();
+    expect(serverFactory.startServer).toHaveBeenCalledWith('/tmp/css.sock');
+  });
+
   it('can stop the server.', async(): Promise<void> => {
+    initializer = new ServerInitializer(serverFactory, 3000);
     await initializer.handle();
     await expect(initializer.finalize()).resolves.toBeUndefined();
     expect(server.close).toHaveBeenCalledTimes(1);
   });
 
   it('only tries to stop the server if it was initialized.', async(): Promise<void> => {
+    initializer = new ServerInitializer(serverFactory, 3000);
     await expect(initializer.finalize()).resolves.toBeUndefined();
     expect(server.close).toHaveBeenCalledTimes(0);
   });